+++ examples/multilingual.rs
@@ -94,7 +94,7 @@ impl MyApp {
 
 impl eframe::App for MyApp {
     fn ui(&mut self, ui: &mut egui::Ui, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show_inside(ui, |ui| {
+        egui::CentralPanel::default().show(ui, |ui| {
             let language_before = self.language.clone();
 
             egui::ComboBox::from_label("Language")